% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-movement.R
\name{.addEventMarkers}
\alias{.addEventMarkers}
\title{Add event markers to plot}
\usage{
.addEventMarkers(p, events, time_axis, schema = NULL)
}
\description{
Add event markers to plot
}
\keyword{internal}
