% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-movement.R
\name{.addPhaseRegions}
\alias{.addPhaseRegions}
\title{Add phase regions to plot}
\usage{
.addPhaseRegions(p, schema, time_axis, colors = NULL)
}
\description{
Add phase regions to plot
}
\keyword{internal}
