% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-c3d.R
\name{.c3d_point_units}
\alias{.c3d_point_units}
\title{Determine the POINT units for a C3D write}
\usage{
.c3d_point_units(x)
}
\arguments{
\item{x}{A PhysioExperiment.}
}
\value{
A single character string (defaults to \code{"mm"}).
}
\description{
Determine the POINT units for a C3D write
}
\keyword{internal}
