% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-events.R
\name{.calculateProminence}
\alias{.calculateProminence}
\title{Calculate peak prominence}
\usage{
.calculateProminence(signal, peaks, type = "max")
}
\description{
Calculate peak prominence
}
\keyword{internal}
