% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-integration.R
\name{.create_gait_schema_opensim}
\alias{.create_gait_schema_opensim}
\title{Create gait schema based on available OpenSim data}
\usage{
.create_gait_schema_opensim(
  has_grf,
  has_heel_marker,
  has_toe_marker,
  has_kinematics,
  available_signals,
  side
)
}
\description{
Create gait schema based on available OpenSim data
}
\keyword{internal}
