% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-integration.R
\name{.create_jump_schema_opensim}
\alias{.create_jump_schema_opensim}
\title{Create jump schema based on available OpenSim data}
\usage{
.create_jump_schema_opensim(has_grf, has_kinematics, available_signals)
}
\description{
Create jump schema based on available OpenSim data
}
\keyword{internal}
