% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-integration.R
\name{.create_running_schema_opensim}
\alias{.create_running_schema_opensim}
\title{Create running schema based on available OpenSim data}
\usage{
.create_running_schema_opensim(
  has_grf,
  has_kinematics,
  available_signals,
  side
)
}
\description{
Create running schema based on available OpenSim data
}
\keyword{internal}
