% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-events.R
\name{.detectByMethod}
\alias{.detectByMethod}
\title{Detect event by specified method}
\usage{
.detectByMethod(evt, signals, data, n_samples, sr, ...)
}
\description{
Detect event by specified method
}
\keyword{internal}
