% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-events.R
\name{.detectSingleEvent}
\alias{.detectSingleEvent}
\title{Detect a single event}
\usage{
.detectSingleEvent(evt, signals, data, n_samples, sr, method, ...)
}
\description{
Detect a single event
}
\keyword{internal}
