% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-movement.R
\name{.extractPlotData}
\alias{.extractPlotData}
\title{Extract plot data from various input types}
\usage{
.extractPlotData(x, channel = 1)
}
\description{
Extract plot data from various input types
}
\keyword{internal}
