% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-integration.R
\name{.looks_like_physioopensim_summary}
\alias{.looks_like_physioopensim_summary}
\title{Check whether object has PhysioOpenSim model summary shape}
\usage{
.looks_like_physioopensim_summary(x)
}
\description{
Check whether object has PhysioOpenSim model summary shape
}
\keyword{internal}
