% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-normalize.R
\name{.normalizeCycle}
\alias{.normalizeCycle}
\title{Cycle normalization}
\usage{
.normalizeCycle(x, norm_length, ...)
}
\arguments{
\item{x}{Data (matrix, PhysioExperiment, or list of matrices)}

\item{norm_length}{Target length}

\item{...}{Additional arguments (unused)}
}
\value{
Normalized data
}
\description{
Normalizes data to a fixed number of points using linear interpolation.
}
\keyword{internal}
