% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-normalize.R
\name{.normalizeDTW}
\alias{.normalizeDTW}
\title{DTW-based normalization}
\usage{
.normalizeDTW(x, norm_length, reference = NULL, ...)
}
\arguments{
\item{x}{Data (list of matrices for multiple trials)}

\item{norm_length}{Target length}

\item{reference}{Reference trial or "mean"}

\item{...}{Additional arguments passed to dtwDistance}
}
\value{
Normalized data
}
\description{
Aligns data using Dynamic Time Warping.
}
\keyword{internal}
