% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-normalize.R
\name{.normalizeLandmark}
\alias{.normalizeLandmark}
\title{Landmark-based normalization}
\usage{
.normalizeLandmark(
  x,
  norm_length,
  events = NULL,
  schema = NULL,
  landmark_event = NULL,
  window_before = NULL,
  window_after = NULL,
  ...
)
}
\arguments{
\item{x}{Data}

\item{norm_length}{Target length}

\item{events}{detected_events object}

\item{schema}{TaskSchema object}

\item{landmark_event}{Name of event to align to}

\item{window_before}{Samples/percent before landmark}

\item{window_after}{Samples/percent after landmark}

\item{...}{Additional arguments}
}
\value{
Normalized data
}
\description{
Aligns data to a specific event (landmark).
}
\keyword{internal}
