% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-normalize.R
\name{.normalizeMatrix}
\alias{.normalizeMatrix}
\title{Normalize a single matrix to fixed length}
\usage{
.normalizeMatrix(mat, norm_length)
}
\description{
Normalize a single matrix to fixed length
}
\keyword{internal}
