% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-normalize.R
\name{.normalizePhase}
\alias{.normalizePhase}
\title{Phase-based normalization}
\usage{
.normalizePhase(x, norm_length, schema = NULL, events = NULL, ...)
}
\arguments{
\item{x}{Data (segmented_phases preferred)}

\item{norm_length}{Target length per phase}

\item{schema}{TaskSchema object}

\item{events}{detected_events object}

\item{...}{Additional arguments}
}
\value{
Normalized data
}
\description{
Normalizes each phase independently to the same length.
}
\keyword{internal}
