% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-normalize.R
\name{.normalizeVector}
\alias{.normalizeVector}
\title{Normalize a single vector to fixed length}
\usage{
.normalizeVector(vec, norm_length)
}
\description{
Normalize a single vector to fixed length
}
\keyword{internal}
