% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-opencap.R
\name{.opencap_get_json}
\alias{.opencap_get_json}
\title{Perform a GET Request to the OpenCap API and Parse JSON}
\usage{
.opencap_get_json(url, auth_header)
}
\arguments{
\item{url}{Character string with the full URL.}

\item{auth_header}{Named character vector with the Authorization header.}
}
\value{
Parsed JSON response as a list.
}
\description{
Perform a GET Request to the OpenCap API and Parse JSON
}
\keyword{internal}
