% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-integration.R
\name{.resolve_opensim_model_input}
\alias{.resolve_opensim_model_input}
\title{Resolve OpenSim model input into marker metadata}
\usage{
.resolve_opensim_model_input(model)
}
\description{
Resolve OpenSim model input into marker metadata
}
\keyword{internal}
