% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-phases.R
\name{.segmentPhaseList}
\alias{.segmentPhaseList}
\title{Segment a list of phases}
\usage{
.segmentPhaseList(phase_list, data, event_idx, include_subphases, sr)
}
\description{
Segment a list of phases
}
\keyword{internal}
