% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-TaskSchema.R
\name{getEvent}
\alias{getEvent}
\title{Get an event by name from a TaskSchema}
\usage{
getEvent(schema, event_name)
}
\arguments{
\item{schema}{A TaskSchema object}

\item{event_name}{Name of the event to retrieve}
}
\value{
An Event object or NULL if not found
}
\description{
Get an event by name from a TaskSchema
}
\examples{
getEvent(schema_gait, getEventNames(schema_gait)[1])
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=getEventNames]{getEventNames()}}, \code{\link[=getPhase]{getPhase()}}, \code{\link[=TaskSchema]{TaskSchema()}}
}
