% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-TaskSchema.R
\name{getEventNames}
\alias{getEventNames}
\title{Get event names from a TaskSchema}
\usage{
getEventNames(schema)
}
\arguments{
\item{schema}{A TaskSchema object}
}
\value{
Character vector of event names
}
\description{
Get event names from a TaskSchema
}
\examples{
getEventNames(schema_gait)
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=getPhaseNames]{getPhaseNames()}}, \code{\link[=getEvent]{getEvent()}}, \code{\link[=TaskSchema]{TaskSchema()}}
}
