% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-TaskSchema.R
\name{getPhaseColors}
\alias{getPhaseColors}
\title{Get phase colors from a TaskSchema}
\usage{
getPhaseColors(schema, include_subphases = FALSE)
}
\arguments{
\item{schema}{A TaskSchema object}

\item{include_subphases}{Whether to include subphase colors}
}
\value{
Named character vector of colors (phase name -> color)
}
\description{
Get phase colors from a TaskSchema
}
\examples{
getPhaseColors(schema_gait)
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=getPhaseNames]{getPhaseNames()}}, \code{\link[=getPhase]{getPhase()}}, \code{\link[=TaskSchema]{TaskSchema()}}
}
