% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality-smoothness.R
\name{ldlj}
\alias{ldlj}
\title{Log Dimensionless Jerk (LDLJ) movement smoothness}
\usage{
ldlj(speed, fs)
}
\arguments{
\item{speed}{Numeric speed profile.}

\item{fs}{Sampling frequency in Hz.}
}
\value{
The LDLJ (negative scalar).
}
\description{
The log of the (speed-based) dimensionless jerk, negated so that more
negative = less smooth.
}
\examples{
t <- seq(0, 1, length.out = 200); tau <- t
speed <- 30 * tau^2 - 60 * tau^3 + 30 * tau^4
ldlj(speed, fs = 200)
}
\references{
Balasubramanian et al. (2015); Melendez-Calderon et al. (2021).
}
