% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical-mobility-tests.R
\name{plotTUG}
\alias{plotTUG}
\title{Plot an instrumented TUG timeline}
\usage{
plotTUG(report, title = "Instrumented TUG")
}
\arguments{
\item{report}{An \code{itug_report} from \code{\link[=instrumentedTUG]{instrumentedTUG()}}.}

\item{title}{Plot title.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots the trunk yaw angular velocity with the detected turns shaded and the
sit-to-stand / turn-to-sit transitions marked.
}
\examples{
fs <- 100
t <- (0:(13 * fs - 1)) / fs
g <- function(t, mu, s) exp(-((t - mu) / s)^2)
yaw <- (pi / (0.6 * sqrt(pi))) * g(t, 5, 0.6) +
       (pi / (0.5 * sqrt(pi))) * g(t, 9.75, 0.5)
pitch <- 1.5 * g(t, 1, 0.3) + 1.3 * g(t, 11.5, 0.3)
av <- cbind(0.05 * sin(2 * pi * t), pitch, yaw)
plotTUG(instrumentedTUG(av, fs))
}
\seealso{
\code{\link[=instrumentedTUG]{instrumentedTUG()}}
}
