% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-fda.R
\name{print.fpca_result}
\alias{print.fpca_result}
\title{Print fPCA result}
\usage{
\method{print}{fpca_result}(x, ...)
}
\arguments{
\item{x}{An fpca_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print fPCA result
}
\references{
Ramsay JO, Silverman BW (2005). "Functional Data Analysis." 2nd ed. Springer.
}
\seealso{
\code{\link[=fPCA]{fPCA()}} for performing functional principal component analysis,
\code{\link[=plotFPCA]{plotFPCA()}} for visualization of fPCA results,
\code{\link[=reconstructFPCA]{reconstructFPCA()}} for waveform reconstruction from fPCA scores.
}
