% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-phases.R
\name{print.segmented_phases}
\alias{print.segmented_phases}
\title{Print method for segmented phases}
\usage{
\method{print}{segmented_phases}(x, ...)
}
\arguments{
\item{x}{A segmented_phases object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for segmented phases
}
\examples{
set.seed(123)
data <- matrix(rnorm(200), 100, 2)
events <- data.frame(
  event = c("hs1","ff","ms","ho","to","hs2"),
  label = c("Heel Strike","Foot Flat","Midstance","Heel Off","Toe Off","Heel Strike 2"),
  index = c(1, 13, 31, 41, 61, 100), time = c(0, .12, .30, .40, .60, 1.0),
  percent = c(0, 12, 30, 40, 60, 100), method = rep("manual", 6),
  confidence = rep(1, 6), stringsAsFactors = FALSE)
class(events) <- c("detected_events", "data.frame")
attr(events, "sampling_rate") <- 100
print(segmentPhases(data, events, schema_gait))
}
