% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas-prebuilt.R
\docType{data}
\name{schema_balance}
\alias{schema_balance}
\title{Pre-built Balance Schema}
\format{
A TaskSchema object for continuous balance assessment.
}
\usage{
schema_balance
}
\description{
Task schema for postural control/balance analysis.
}
\examples{
print(schema_balance)

}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=TaskSchema]{TaskSchema()}}, \link{schema_gait}, \link{schema_cutting}
}
\keyword{datasets}
