% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas-prebuilt.R
\docType{data}
\name{schema_throw}
\alias{schema_throw}
\title{Pre-built Throwing Schema}
\format{
A TaskSchema object with throwing-specific events and phases.
}
\usage{
schema_throw
}
\description{
Task schema for throwing motion analysis (baseball, softball, etc.).
}
\examples{
print(schema_throw)

}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=TaskSchema]{TaskSchema()}}, \link{schema_gait}, \link{schema_cutting}
}
\keyword{datasets}
