% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement-tremor.R
\name{tremorAmplitude}
\alias{tremorAmplitude}
\title{Band-limited tremor amplitude (RMS)}
\usage{
tremorAmplitude(signal, sampling_rate, band = c(3, 12), order = 4)
}
\arguments{
\item{signal}{Numeric movement signal.}

\item{sampling_rate}{Sampling rate in Hz.}

\item{band}{Length-2 tremor band \code{c(low, high)} in Hz (default \code{c(3, 12)}).}

\item{order}{Butterworth filter order (default 4).}
}
\value{
A list with \code{rms} and the \code{band} used.
}
\description{
Band-pass filters the signal to the tremor band and returns its root-mean-
square amplitude (in the units of the input signal).
}
\examples{
fs <- 100; t <- seq(0, 20, by = 1 / fs)
tremorAmplitude(2 * sin(2 * pi * 5 * t), fs)$rms   # ~ 2 / sqrt(2)
}
\seealso{
\code{\link[=tremorSpectrum]{tremorSpectrum()}}, \code{\link[=tremorMetrics]{tremorMetrics()}}
}
