% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-TaskSchema.R
\name{validateSchema}
\alias{validateSchema}
\title{Validate a TaskSchema}
\usage{
validateSchema(schema)
}
\arguments{
\item{schema}{A TaskSchema object}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
Checks that a TaskSchema is internally consistent.
}
\examples{
validateSchema(schema_gait)
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=TaskSchema]{TaskSchema()}}, \code{\link[=getSchema]{getSchema()}}, \code{\link[=listSchemas]{listSchemas()}}
}
