% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality-smoothness.R
\name{dimensionlessJerk}
\alias{dimensionlessJerk}
\title{Dimensionless jerk of a speed profile}
\usage{
dimensionlessJerk(speed, fs)
}
\arguments{
\item{speed}{Numeric speed profile.}

\item{fs}{Sampling frequency in Hz.}
}
\value{
The (positive) dimensionless jerk. See \code{\link{ldlj}} for the log form.
}
\description{
Dimensionless jerk of a speed profile
}
