% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-phases.R
\name{extractPhase}
\alias{extractPhase}
\title{Extract a single phase from segmented data}
\usage{
extractPhase(x, phase_name, search_subphases = TRUE)
}
\arguments{
\item{x}{A segmented_phases object}

\item{phase_name}{Name of the phase to extract}

\item{search_subphases}{Whether to search within subphases}
}
\value{
A matrix of phase data
}
\description{
Extract a single phase from segmented data
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=segmentPhases]{segmentPhases()}}, \code{\link[=phaseTiming]{phaseTiming()}}, \code{\link[=getPhaseData]{getPhaseData()}}
}
