% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-TaskSchema.R
\name{getPhase}
\alias{getPhase}
\title{Get a phase by name from a TaskSchema}
\usage{
getPhase(schema, phase_name, search_subphases = TRUE)
}
\arguments{
\item{schema}{A TaskSchema object}

\item{phase_name}{Name of the phase to retrieve}

\item{search_subphases}{Whether to search within subphases}
}
\value{
A Phase object or NULL if not found
}
\description{
Get a phase by name from a TaskSchema
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=getPhaseNames]{getPhaseNames()}}, \code{\link[=getPhaseColors]{getPhaseColors()}}, \code{\link[=getEvent]{getEvent()}}
}
