% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-phases.R
\name{getPhaseData}
\alias{getPhaseData}
\title{Get data for all phases as a list of matrices}
\usage{
getPhaseData(x, include_subphases = FALSE)
}
\arguments{
\item{x}{A segmented_phases object}

\item{include_subphases}{Whether to include subphases}
}
\value{
Named list of matrices
}
\description{
Get data for all phases as a list of matrices
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=segmentPhases]{segmentPhases()}}, \code{\link[=extractPhase]{extractPhase()}}, \code{\link[=hasValidPhases]{hasValidPhases()}}
}
