% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-TaskSchema.R
\name{getPhaseNames}
\alias{getPhaseNames}
\title{Get phase names from a TaskSchema}
\usage{
getPhaseNames(schema, include_subphases = FALSE)
}
\arguments{
\item{schema}{A TaskSchema object}

\item{include_subphases}{Whether to include subphase names}
}
\value{
Character vector of phase names
}
\description{
Get phase names from a TaskSchema
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=getEventNames]{getEventNames()}}, \code{\link[=getPhase]{getPhase()}}, \code{\link[=getPhaseColors]{getPhaseColors()}}
}
