% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-phases.R
\name{hasValidPhases}
\alias{hasValidPhases}
\title{Check if all phases are valid}
\usage{
hasValidPhases(x)
}
\arguments{
\item{x}{A segmented_phases object}
}
\value{
Logical indicating if all phases have valid data
}
\description{
Check if all phases are valid
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=segmentPhases]{segmentPhases()}}, \code{\link[=getPhaseData]{getPhaseData()}}, \code{\link[=extractPhase]{extractPhase()}}
}
