% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-phases.R
\name{phaseDurations}
\alias{phaseDurations}
\title{Get phase durations}
\usage{
phaseDurations(x, unit = c("percent", "seconds", "samples"))
}
\arguments{
\item{x}{A segmented_phases object}

\item{unit}{"samples", "seconds", or "percent"}
}
\value{
Named numeric vector of phase durations
}
\description{
Get phase durations
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=phaseTiming]{phaseTiming()}}, \code{\link[=phaseRatios]{phaseRatios()}}, \code{\link[=segmentPhases]{segmentPhases()}}
}
