% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-phases.R
\name{phaseRatios}
\alias{phaseRatios}
\title{Calculate phase ratios}
\usage{
phaseRatios(x, reference = "total")
}
\arguments{
\item{x}{A segmented_phases object}

\item{reference}{Reference for ratio calculation ("total" or phase name)}
}
\value{
Named numeric vector of phase ratios
}
\description{
Calculate phase ratios
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=phaseDurations]{phaseDurations()}}, \code{\link[=phaseTiming]{phaseTiming()}}, \code{\link[=segmentPhases]{segmentPhases()}}
}
