% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-phases.R
\name{phaseTiming}
\alias{phaseTiming}
\title{Get phase timing information}
\usage{
phaseTiming(x, as_percent = TRUE)
}
\arguments{
\item{x}{A segmented_phases object}

\item{as_percent}{Return timing as percentage (TRUE) or seconds (FALSE)}
}
\value{
A data.frame with phase timing information
}
\description{
Get phase timing information
}
\references{
Winter DA (2009). "Biomechanics and Motor Control of Human Movement."
4th ed. Wiley.
}
\seealso{
\code{\link[=segmentPhases]{segmentPhases()}}, \code{\link[=phaseDurations]{phaseDurations()}}, \code{\link[=phaseRatios]{phaseRatios()}}
}
