% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-dimred.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Plot UMAP embedding}
\usage{
plotUMAP(x, groups = NULL, labels = NULL)
}
\arguments{
\item{x}{A waveform_umap object.}

\item{groups}{Optional grouping factor.}

\item{labels}{Optional point labels.}
}
\value{
A ggplot object.
}
\description{
Creates a scatter plot of UMAP embedding.
}
\references{
van der Maaten L, Hinton G (2008). "Visualizing Data using t-SNE."
Journal of Machine Learning Research, 9, 2579-2605.
}
\seealso{
\code{\link[=waveformUMAP]{waveformUMAP()}}, \code{\link[=plotPCAScatter]{plotPCAScatter()}}, \code{\link[=waveformPCA]{waveformPCA()}}
}
