% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-events.R
\name{print.detected_events}
\alias{print.detected_events}
\title{Print method for detected events}
\usage{
\method{print}{detected_events}(x, ...)
}
\arguments{
\item{x}{A detected_events object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for detected events
}
