% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-dtw.R
\name{print.dtw_clustering}
\alias{print.dtw_clustering}
\title{Print DTW clustering result}
\usage{
\method{print}{dtw_clustering}(x, ...)
}
\arguments{
\item{x}{A dtw_clustering object}

\item{...}{Additional arguments (unused)}
}
\description{
Print DTW clustering result
}
