% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-dtw.R
\name{print.dtw_result}
\alias{print.dtw_result}
\title{Print DTW result}
\usage{
\method{print}{dtw_result}(x, ...)
}
\arguments{
\item{x}{A dtw_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print DTW result
}
