% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-phases.R
\name{print.multi_trial_phases}
\alias{print.multi_trial_phases}
\title{Print method for multi-trial phases}
\usage{
\method{print}{multi_trial_phases}(x, ...)
}
\arguments{
\item{x}{A multi_trial_phases object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for multi-trial phases
}
