% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-phases.R
\name{print.segmented_phases}
\alias{print.segmented_phases}
\title{Print method for segmented phases}
\usage{
\method{print}{segmented_phases}(x, ...)
}
\arguments{
\item{x}{A segmented_phases object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for segmented phases
}
