% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocap-skeleton.R
\name{print.SkeletonModel}
\alias{print.SkeletonModel}
\title{Print a SkeletonModel object}
\usage{
\method{print}{SkeletonModel}(x, ...)
}
\arguments{
\item{x}{A \code{SkeletonModel} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a SkeletonModel object
}
