% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-dimred.R
\name{print.waveform_pca}
\alias{print.waveform_pca}
\title{Print waveform PCA result}
\usage{
\method{print}{waveform_pca}(x, ...)
}
\arguments{
\item{x}{A waveform_pca object}

\item{...}{Additional arguments (unused)}
}
\description{
Print waveform PCA result
}
