% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-dimred.R
\name{print.waveform_umap}
\alias{print.waveform_umap}
\title{Print waveform UMAP result}
\usage{
\method{print}{waveform_umap}(x, ...)
}
\arguments{
\item{x}{A waveform_umap object}

\item{...}{Additional arguments (unused)}
}
\description{
Print waveform UMAP result
}
