% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-clinical.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{icc}
\alias{sem}
\alias{mdc}
\alias{cohensD}
\alias{etaSquared}
\alias{blandAltman}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{PhysioCore}{\code{\link[PhysioCore]{blandAltman}}, \code{\link[PhysioCore]{cohensD}}, \code{\link[PhysioCore]{etaSquared}}, \code{\link[PhysioCore]{icc}}, \code{\link[PhysioCore]{mdc}}, \code{\link[PhysioCore]{sem}}}
}}

