% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-fda.R, R/stats-clinical.R,
%   R/stats-reliability.R, R/stats-circular.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fPCA}
\alias{reconstructFPCA}
\alias{registerCurves}
\alias{icc}
\alias{sem}
\alias{mdc}
\alias{cohensD}
\alias{etaSquared}
\alias{blandAltman}
\alias{waveformCMC}
\alias{waveformICC}
\alias{waveformReliability}
\alias{circularSummary}
\alias{rayleighTest}
\alias{watsonWilliamsTest}
\alias{circularLinearCorrelation}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{PhysioCore}{\code{\link[PhysioCore]{blandAltman}}, \code{\link[PhysioCore]{circularLinearCorrelation}}, \code{\link[PhysioCore]{circularSummary}}, \code{\link[PhysioCore]{cohensD}}, \code{\link[PhysioCore]{etaSquared}}, \code{\link[PhysioCore]{fPCA}}, \code{\link[PhysioCore]{icc}}, \code{\link[PhysioCore]{mdc}}, \code{\link[PhysioCore]{rayleighTest}}, \code{\link[PhysioCore]{reconstructFPCA}}, \code{\link[PhysioCore]{registerCurves}}, \code{\link[PhysioCore]{sem}}, \code{\link[PhysioCore]{watsonWilliamsTest}}, \code{\link[PhysioCore]{waveformCMC}}, \code{\link[PhysioCore]{waveformICC}}, \code{\link[PhysioCore]{waveformReliability}}}
}}

